Safari-611.1.10/LayoutTests/editing/pasteboard/dom-paste/dom-paste-confirmation-expected.txt

Click here to copy
Click here to copy

Verifies that a menu is shown when the page programmatically triggers paste, and that selecting Paste in the menu allows the paste to happen. To manually test, click or tap the text on the bottom, click or tap the editable area above, and then select 'Paste' in the resulting menu. The text 'Click here to copy' should be pasted twice in the editable area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.queryCommandSupported('Paste') is true
PASS document.queryCommandEnabled('Paste') is true
PASS event.clipboardData.getData('text/plain') is "Click here to copy"
PASS document.execCommand('Paste') is true
PASS event.clipboardData.getData('text/plain') is "Click here to copy"
PASS document.execCommand('Paste') is true
PASS editor.textContent is "Click here to copyClick here to copy"
PASS successfullyParsed is true

TEST COMPLETE
